pyproject.toml
[tool.poetry]
name = "campsites"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fake-useragent = "^0.1.11"
requests = "^2.25.1"
click = "^7.1.2"
twilio = "^6.55.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.poetry.scripts]
find-campsites = 'campsites.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
